Fix memory amount

+++ src/main/scala/pw/ian/sysadmincraft/system/ProcessAdmin.scala
@@ -8,7 +8,7 @@ case class SysProcess(ids: Set[Int], realMemory: Long, virtualMemory: Long, name
 
   def totalMemory = realMemory + virtualMemory
 
-  def memAmt: Double = Math.min(totalMemory / WorldConstants.MAX_MEMORY, 1)
+  def memAmt: Double = Math.min(totalMemory.toDouble / WorldConstants.MAX_MEMORY, 1)
 
   def kill() = {
     // ids.foreach(x => s"kill -9 $x" !)
